services/users/requirements.txt
Flask==1.0.2
flasgger==0.9.0
Flask-SQLAlchemy==2.3.2
psycopg2==2.7.4
Flask-Testing==0.6.2
coverage==4.5.1
flake8==3.5.0
